Back_end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
bcrypt = "^4.1.3"
google-generativeai = "^0.5.4"
psycopg2 = "^2.9.9"
pyjwt = "^2.8.0"
pyjst = "^0.1.6"
jwt = "^1.3.1"
confluent-kafka = "^2.4.0"
aiokafka = "^0.10.0"
asyncio = "^3.4.3"
sqlmodel = "^0.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
